requirements.txt
azure-cognitiveservices-vision-face
msrest
streamlit
